+++ packages/dicebear/src/commands/build.ts
@@ -1,4 +1,0 @@
-import { Command } from 'commander';
-import { build as buildAction } from '@dicebear/build';
-
-export const build = new Command('build').arguments('<name>').action(buildAction);